Add files via upload
In decision-making, we face a situation where we have to decide whether it is worth performing an investment or not, based on a multitude of variables. The Net Present Value (NPV) indicates the profit that our investment will produce over a series of time periods within a fixed time horizon. Its computation begins with certain data or variables from which we can obtain a cash-flow for each time period under consideration. This study will focus on the variables involved in the aforementioned computation that can take values within an interval of possibilities whose end-points are their pessimistic forecast (PF) and their optimistic one (OF). Now, we are in front of infinite different possible forecasts, all those values between the end-points, and the expert in charge needs to determine which data is more appropriate among them in order to make the final decision. The objective of this study is to introduce two analytical techniques based on the use of aggregation functions and mid-points to select one forecast among the aforesaid infinite range. These new methodologies incorporate a penalization, fixed following an analytical procedure and thus reducing subjectivity, for the elapsed time and for the discrepancy between PF and OF in order to generate the NPV. In contrast, the classical methodology, based on the Certainty Equivalent, operates under the estimation of some reduction coefficients fixed subjectively by the expert for penalizing those forecasts of the cash-flow corresponding to more distant periods. Finally, all methodologies are tested by applying them to a paradigmatic example where real data is considered and the NPV of a five-year hotel assessment is computed. Here, it is illustrated that the proposed methodologies could be appropriate for the assessment of investments and companies.